Replaced canvas drawer with svg to canvas drawer
